refactor: extract bootstrap_stdlib to break up blob anti-pattern
